begin shifting endpoint /form-letters to api folder
