+++ crt_portal/api/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 from rest_framework.urlpatterns import format_suffix_patterns
-from api.views import ResponseList, ResponseDetail, ReportCountView, ReportList, ReportDetail, api_root
+from api.views import ResponseList, ResponseDetail, ReportCountView, ReportList, ReportDetail, FormLettersIndex, api_root
 
 app_name = 'api'
 
@@ -11,6 +11,7 @@ urlpatterns = [
     path('responses/', ResponseList.as_view(), name='response-list'),
     path('responses/<int:pk>/', ResponseDetail.as_view(), name='response-detail'),
     path('report-count/', ReportCountView.as_view(), name='report-count'),
+    path('form-letters/', FormLettersIndex.as_view(), name='form-letters')
 ]
 
 urlpatterns = format_suffix_patterns(urlpatterns)
