--- crt_portal/cts_forms/templates/form_letters.html
@@ -1,8 +1,0 @@
-{% load static %}
-{% block content %}
-  <h1>Form Letters Sent</h1>
-
-
-  <p>Total Number of Letters Sent: {{ total_form_letters_sent }} </p>
-  <p>Form Letters Dictionary: {{ emails_counter }} </p>
-{% endblock %}
